+++ asn1/api/lc_x509_common.h
@@ -257,7 +257,7 @@ struct lc_public_key_signature {
 
 struct lc_x509_certificate_name_component {
 	const char *value;
-	uint8_t size;
+	size_t size;
 };
 
 struct lc_x509_certificate_name {
@@ -346,9 +346,6 @@ struct lc_x509_certificate {
 	size_t raw_akid_size;
 	const uint8_t *raw_akid; /* authority key Id binary format */
 	unsigned int index;
-	char issuer[LC_ASN1_MAX_ISSUER_NAME + 1]; /* Name of certificate issuer */
-	char subject[LC_ASN1_MAX_ISSUER_NAME +
-		     1]; /* Name of certificate subject */
 
 	uint8_t x509_version; /* X.509 Version of certificate */
 	unsigned int seen : 1; /* Infinite recursion prevention */